src/lambdas/timesheet_etl/requirements.txt
anthropic>=0.40.0
psycopg2-binary>=2.9.9
pypdf>=4.0.0
